+++ src/views/concours/editions/components/modal/StatsModal.vue
